Reconnect on failure to resolve gateway addresses instead of entering error state
