+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/states/connecting_state.rs
@@ -3,6 +3,7 @@
 
 #[cfg(any(target_os = "linux", target_os = "macos", target_os = "windows"))]
 use std::net::SocketAddr;
+use std::time::Duration;
 
 use futures::{
     FutureExt,
@@ -36,11 +37,21 @@ use crate::tunnel_state_machine::{
     },
 };
 
+/// Initial delay between retry attempts.
+const INITIAL_WAIT_DELAY: Duration = Duration::from_secs(2);
+
+/// Wait delay multiplier used for each subsequent retry attempt.
+const DELAY_MULTIPLIER: u32 = 2;
+
+/// Max wait delay between retry attempts.
+const MAX_WAIT_DELAY: Duration = Duration::from_secs(15);
+
 /// Default websocket port used as a fallback
 #[cfg(any(target_os = "linux", target_os = "macos", target_os = "windows"))]
 const DEFAULT_WS_PORT: u16 = 80;
 
 type ResolveConfigFuture = BoxFuture<'static, Result<ResolvedConfig>>;
+type ReconnectDelayFuture = BoxFuture<'static, ()>;
 
 pub struct ConnectingState {
     retry_attempt: u32,
@@ -50,6 +61,7 @@ pub struct ConnectingState {
     selected_gateways: Option<SelectedGateways>,
     resolved_gateway_config: Option<ResolvedConfig>,
     resolve_config_fut: Fuse<ResolveConfigFuture>,
+    reconnect_delay_fut: Fuse<ReconnectDelayFuture>,
 }
 
 impl ConnectingState {
@@ -98,14 +110,14 @@ impl ConnectingState {
             }
         }
 
-        let gateway_config = shared_state.nym_config.gateway_config.clone();
-        let resolve_config_fut = async move {
-            nym_gateway_directory::resolve_config(&gateway_config)
-                .await
-                .map_err(Error::ResolveGatewayAddrs)
-        }
-        .boxed()
-        .fuse();
+        let resolve_config_fut = Fuse::terminated();
+        let reconnect_delay_fut = if retry_attempt > 0 {
+            let wait_delay = wait_delay(retry_attempt);
+            tracing::info!("Waiting {}s before reconnect", wait_delay.as_secs());
+            tokio::time::sleep(wait_delay).boxed().fuse()
+        } else {
+            std::future::ready(()).boxed().fuse()
+        };
 
         let (monitor_event_sender, monitor_event_receiver) = mpsc::unbounded_channel();
 
@@ -118,9 +130,10 @@ impl ConnectingState {
                 selected_gateways,
                 resolved_gateway_config: None,
                 resolve_config_fut,
+                reconnect_delay_fut,
             }),
             PrivateTunnelState::Connecting {
-                retry_attempt: 0,
+                retry_attempt,
                 connection_data: None,
             },
         )
@@ -282,13 +295,13 @@ impl ConnectingState {
                 resolved_gateway_config
             }
             Err(e) => {
+                let next_attempt = self.retry_attempt.saturating_add(1);
+                tracing::info!(
+                    "Failed to resolve gateway config: {e}. Reconnecting, attempt {next_attempt}."
+                );
                 return NextTunnelState::NewState(
-                    ErrorState::enter(
-                        e.error_state_reason()
-                            .expect("failed to map to error reason"),
-                        shared_state,
-                    )
-                    .await,
+                    ConnectingState::enter(next_attempt, self.selected_gateways, shared_state)
+                        .await,
                 );
             }
         };
@@ -363,7 +376,6 @@ impl ConnectingState {
             resolved_gateway_config: resolved_gateway_config.clone(),
             tunnel_settings: shared_state.tunnel_settings.clone(),
             selected_gateways: self.selected_gateways.clone(),
-            retry_attempt: self.retry_attempt,
         };
         let tunnel_monitor_handle = TunnelMonitor::start(
             tunnel_parameters,
@@ -457,114 +469,124 @@ impl TunnelStateHandler for ConnectingState {
         shared_state: &'async_trait mut SharedState,
     ) -> NextTunnelState {
         tokio::select! {
+            _ = &mut self.reconnect_delay_fut => {
+                let gateway_config = shared_state.nym_config.gateway_config.clone();
+
+                self.resolve_config_fut = async move {
+                    nym_gateway_directory::resolve_config(&gateway_config)
+                        .await
+                        .map_err(Error::ResolveApiHostnames)
+                }
+                .boxed()
+                .fuse();
+
+                NextTunnelState::SameState(self)
+            },
             resolved_gateway_config = &mut self.resolve_config_fut => {
                 self.handle_resolved_gateway_config(resolved_gateway_config, shared_state).await
             }
             Some(monitor_event) = self.tunnel_monitor_event_receiver.recv() => {
-            match monitor_event {
-                TunnelMonitorEvent::ReconnectCooldown { .. } => {
-                    NextTunnelState::SameState(self)
-                }
-                TunnelMonitorEvent::InitializingClient => {
-                    NextTunnelState::SameState(self)
-                }
-                TunnelMonitorEvent::SelectingGateways => {
-                    NextTunnelState::SameState(self)
-                }
-                TunnelMonitorEvent::SelectedGateways {
-                    gateways, reply_tx
-                } => {
-                   let next_state = match self.handle_selected_gateways(gateways, shared_state).await {
-                        Ok(()) => {
-                            NextTunnelState::SameState(self)
-                        }
-                        Err(e) => {
-                            if let Some(tunnel_monitor_handle) = self.tunnel_monitor_handle {
-                                NextTunnelState::NewState(DisconnectingState::enter(
-                                    // todo: fix that expect()
-                                    PrivateActionAfterDisconnect::Error(e.error_state_reason().expect("failed to obtain error state reason")),
-                                    tunnel_monitor_handle,
-                                    shared_state
-                                ))
-                            } else {
-                                NextTunnelState::NewState(ErrorState::enter(
-                                    // todo: fix that expect()
-                                    e.error_state_reason().expect("failed to obtain error state reason"),
-                                    shared_state
-                                ).await)
+                match monitor_event {
+                    TunnelMonitorEvent::InitializingClient => {
+                        NextTunnelState::SameState(self)
+                    }
+                    TunnelMonitorEvent::SelectingGateways => {
+                        NextTunnelState::SameState(self)
+                    }
+                    TunnelMonitorEvent::SelectedGateways {
+                        gateways, reply_tx
+                    } => {
+                    let next_state = match self.handle_selected_gateways(gateways, shared_state).await {
+                            Ok(()) => {
+                                NextTunnelState::SameState(self)
                             }
-                        }
-                    };
-                    _ = reply_tx.send(());
-                    next_state
-                }
-                TunnelMonitorEvent::InterfaceUp {
-                    tunnel_interface, connection_data, reply_tx
-                }  => {
-                    let next_state = match self.handle_interface_up(tunnel_interface, shared_state).await {
-                        Ok(()) => {
-                            let state = PrivateTunnelState::Connecting { retry_attempt: self.retry_attempt, connection_data: Some(*connection_data) };
-                            NextTunnelState::NewState((self, state))
-                        },
-                        Err(e) => {
-                            if let Some(tunnel_monitor_handle) = self.tunnel_monitor_handle {
-                                NextTunnelState::NewState(DisconnectingState::enter(
-                                    // todo: fix that expect()
-                                    PrivateActionAfterDisconnect::Error(e.error_state_reason().expect("failed to obtain error state reason")),
-                                    tunnel_monitor_handle,
-                                    shared_state
-                                ))
-                            } else {
-                                NextTunnelState::NewState(ErrorState::enter(
-                                    // todo: fix that expect()
-                                    e.error_state_reason().expect("failed to obtain error state reason"),
-                                    shared_state
-                                ).await)
+                            Err(e) => {
+                                if let Some(tunnel_monitor_handle) = self.tunnel_monitor_handle {
+                                    NextTunnelState::NewState(DisconnectingState::enter(
+                                        // todo: fix that expect()
+                                        PrivateActionAfterDisconnect::Error(e.error_state_reason().expect("failed to obtain error state reason")),
+                                        tunnel_monitor_handle,
+                                        shared_state
+                                    ))
+                                } else {
+                                    NextTunnelState::NewState(ErrorState::enter(
+                                        // todo: fix that expect()
+                                        e.error_state_reason().expect("failed to obtain error state reason"),
+                                        shared_state
+                                    ).await)
+                                }
                             }
-                        }
-                    };
-                    _ = reply_tx.send(());
-                    next_state
-                }
-                TunnelMonitorEvent::Up { tunnel_interface, connection_data } => {
-                    NextTunnelState::NewState(ConnectedState::enter(
-                        tunnel_interface,
-                        *connection_data,
-                        self.selected_gateways.expect("selected gateways must be set"),
-                        self.tunnel_monitor_handle.expect("monitor handle must be set!"),
-                        self.tunnel_monitor_event_receiver,
-                        shared_state,
-                    ).await)
-                }
-                TunnelMonitorEvent::Down { error_state_reason, reply_tx } => {
-                    // Signal that the message was received first.
-                    _ = reply_tx.send(());
-
-                    if let Some(error_state_reason) = error_state_reason {
-                        NextTunnelState::NewState(DisconnectingState::enter(
-                            PrivateActionAfterDisconnect::Error(error_state_reason),
+                        };
+                        _ = reply_tx.send(());
+                        next_state
+                    }
+                    TunnelMonitorEvent::InterfaceUp {
+                        tunnel_interface, connection_data, reply_tx
+                    }  => {
+                        let next_state = match self.handle_interface_up(tunnel_interface, shared_state).await {
+                            Ok(()) => {
+                                let state = PrivateTunnelState::Connecting { retry_attempt: self.retry_attempt, connection_data: Some(*connection_data) };
+                                NextTunnelState::NewState((self, state))
+                            },
+                            Err(e) => {
+                                if let Some(tunnel_monitor_handle) = self.tunnel_monitor_handle {
+                                    NextTunnelState::NewState(DisconnectingState::enter(
+                                        // todo: fix that expect()
+                                        PrivateActionAfterDisconnect::Error(e.error_state_reason().expect("failed to obtain error state reason")),
+                                        tunnel_monitor_handle,
+                                        shared_state
+                                    ))
+                                } else {
+                                    NextTunnelState::NewState(ErrorState::enter(
+                                        // todo: fix that expect()
+                                        e.error_state_reason().expect("failed to obtain error state reason"),
+                                        shared_state
+                                    ).await)
+                                }
+                            }
+                        };
+                        _ = reply_tx.send(());
+                        next_state
+                    }
+                    TunnelMonitorEvent::Up { tunnel_interface, connection_data } => {
+                        NextTunnelState::NewState(ConnectedState::enter(
+                            tunnel_interface,
+                            *connection_data,
+                            self.selected_gateways.expect("selected gateways must be set"),
                             self.tunnel_monitor_handle.expect("monitor handle must be set!"),
-                            shared_state
-                        ))
-                    } else {
-                        if let Some(tunnel_monitor_handle) = self.tunnel_monitor_handle {
-                            let tombstone = tunnel_monitor_handle.wait().await;
-                            Self::handle_tunnel_close(tombstone, shared_state).await;
-                        }
-
-                        let next_attempt = self.retry_attempt.saturating_add(1);
-                        tracing::info!(
-                            "Tunnel closed. Reconnecting, attempt {}.",
-                            next_attempt
-                        );
-                        NextTunnelState::NewState(ConnectingState::enter(
-                            next_attempt,
-                            self.selected_gateways,
-                            shared_state
+                            self.tunnel_monitor_event_receiver,
+                            shared_state,
                         ).await)
                     }
+                    TunnelMonitorEvent::Down { error_state_reason, reply_tx } => {
+                        // Signal that the message was received first.
+                        _ = reply_tx.send(());
+
+                        if let Some(error_state_reason) = error_state_reason {
+                            NextTunnelState::NewState(DisconnectingState::enter(
+                                PrivateActionAfterDisconnect::Error(error_state_reason),
+                                self.tunnel_monitor_handle.expect("monitor handle must be set!"),
+                                shared_state
+                            ))
+                        } else {
+                            if let Some(tunnel_monitor_handle) = self.tunnel_monitor_handle {
+                                let tombstone = tunnel_monitor_handle.wait().await;
+                                Self::handle_tunnel_close(tombstone, shared_state).await;
+                            }
+
+                            let next_attempt = self.retry_attempt.saturating_add(1);
+                            tracing::info!(
+                                "Tunnel closed. Reconnecting, attempt {}.",
+                                next_attempt
+                            );
+                            NextTunnelState::NewState(ConnectingState::enter(
+                                next_attempt,
+                                self.selected_gateways,
+                                shared_state
+                            ).await)
+                        }
+                    }
                 }
-            }
            }
             Some(command) = command_rx.recv() => {
                 match command {
@@ -626,4 +648,10 @@ impl TunnelStateHandler for ConnectingState {
             }
         }
     }
+}
+
+fn wait_delay(retry_attempt: u32) -> Duration {
+    let multiplier = retry_attempt.saturating_mul(DELAY_MULTIPLIER);
+    let delay = INITIAL_WAIT_DELAY.saturating_mul(multiplier);
+    std::cmp::min(delay, MAX_WAIT_DELAY)
 }
