+++ nym-vpn-core/crates/nym-vpn-lib/src/tunnel_state_machine/tunnel_monitor.rs
@@ -16,12 +16,7 @@ use std::os::fd::BorrowedFd;
 use std::os::fd::{AsRawFd, IntoRawFd};
 #[cfg(target_os = "android")]
 use std::os::fd::{FromRawFd, OwnedFd};
-use std::{
-    cmp,
-    net::IpAddr,
-    path::PathBuf,
-    time::{Duration, Instant},
-};
+use std::{net::IpAddr, path::PathBuf, time::Duration};
 #[cfg(unix)]
 use std::{os::fd::RawFd, sync::Arc};
 
@@ -109,15 +104,6 @@ const WG_EXIT_WINTUN_GUID: &str = "{AFE43773-E1F8-4EBB-8536-176AB86AFE9C}";
 pub type TunnelMonitorEventSender = mpsc::UnboundedSender<TunnelMonitorEvent>;
 pub type TunnelMonitorEventReceiver = mpsc::UnboundedReceiver<TunnelMonitorEvent>;
 
-/// Initial delay between retry attempts.
-const INITIAL_WAIT_DELAY: Duration = Duration::from_secs(2);
-
-/// Wait delay multiplier used for each subsequent retry attempt.
-const DELAY_MULTIPLIER: u32 = 2;
-
-/// Max wait delay between retry attempts.
-const MAX_WAIT_DELAY: Duration = Duration::from_secs(15);
-
 /// Timeout when waiting for reply from the event handler.
 const REPLY_TIMEOUT: Duration = Duration::from_secs(5);
 
@@ -126,16 +112,6 @@ const TASK_MANAGER_SHUTDOWN_TIMEOUT_SECS: u64 = 10;
 
 #[derive(Debug)]
 pub enum TunnelMonitorEvent {
-    /// Awaiting a cooldown period before reconnecting
-    #[allow(dead_code)]
-    ReconnectCooldown {
-        /// Cooldown begin time
-        begin_time: Instant,
-
-        /// Cooldown duration
-        duration: Duration,
-    },
-
     /// Initializing mixnet client
     InitializingClient,
 
@@ -204,7 +180,6 @@ pub struct TunnelParameters {
     pub resolved_gateway_config: ResolvedConfig,
     pub tunnel_settings: TunnelSettings,
     pub selected_gateways: Option<SelectedGateways>,
-    pub retry_attempt: u32,
 }
 
 pub struct TunnelMonitor {
@@ -289,20 +264,6 @@ impl TunnelMonitor {
     }
 
     async fn run_inner(&mut self, task_manager: &mut TaskManager) -> Result<Tombstone> {
-        if self.tunnel_parameters.retry_attempt > 0 {
-            let delay = wait_delay(self.tunnel_parameters.retry_attempt);
-            tracing::debug!("Waiting for {}s before connecting.", delay.as_secs());
-            self.send_event(TunnelMonitorEvent::ReconnectCooldown {
-                begin_time: Instant::now(),
-                duration: delay,
-            });
-
-            self.shutdown_token
-                .run_until_cancelled(tokio::time::sleep(delay))
-                .await
-                .ok_or(Error::Tunnel(Box::new(tunnel::Error::Cancelled)))?;
-        }
-
         if self.enable_ipv6() && !ipv6_availability::is_ipv6_enabled_in_os().await {
             return Err(Error::Ipv6Unavailable);
         }
@@ -1373,12 +1334,6 @@ impl TunnelMonitor {
     fn enable_ipv6(&self) -> bool {
         self.tunnel_parameters.tunnel_settings.enable_ipv6
     }
-}
-
-fn wait_delay(retry_attempt: u32) -> Duration {
-    let multiplier = retry_attempt.saturating_mul(DELAY_MULTIPLIER);
-    let delay = INITIAL_WAIT_DELAY.saturating_mul(multiplier);
-    cmp::min(delay, MAX_WAIT_DELAY)
 }
 
 pub struct StartTunnelResult {
